refactor: rename FeatureBenchEvaluator to BenchmarkEvaluator
Rebase of #1449 scoped to the evaluator rename only. The inner-loop
half of the original PR (FeatureBenchInnerLoop → BenchmarkInnerLoop)
is dropped because main removed the class entirely in #1458
(Unified Task.run()); re-adding it would regress that refactor.

- git mv featurebench_evaluator.py → benchmark_evaluator.py
  (class FeatureBenchEvaluator → BenchmarkEvaluator)
- Backward-compat shim at old path (assignment alias + __all__)
- Update parse_pytest_stdout imports in inner_loop.py and
  direct_evaluator.py
- Rename test file and test class; test_multi_benchmark_e2e.py
  backward-compat alias test continues to validate the shim
- Update CLAUDE.md, docs/outer-loop.md, codecov.yml
- Also fix stale references to featurebench_inner_loop.py left on
  main by #1458 (CLAUDE.md module list, codecov.yml ignore list)
- DirectFeatureBenchEvaluator and 'featurebench' benchmark name
  string unchanged (genuinely FeatureBench-specific)

Closes #1421
