+++ codecov.yml
@@ -12,4 +12,4 @@ ignore:
   - "eval/"
   - "docs/"
   - "factory/outer_loop/evaluators/"
-  - "factory/outer_loop/featurebench_inner_loop.py"
+  - "factory/outer_loop/featurebench_evaluator.py"
